feat: Excellent commit
In terms of actual code, all I've done is add the file empty.txt to this
repository.  I got rid of the other commits because frankly, you don't
care about how many times I changed my mind about what to name it.  All
you really wanted to know is that I added the file.

Furthermore this is an example of a commit that follows conventional changelog
commit formatting guidelines.  It is an example of version control as
documentation and is a powerful way to communicate your thoughts and
rationale to other developers. Because I've created a demo issue on our
eng-talks repository located at
https://github.com/nextgxdx/eng-talks/issues/1 you can see that I a
referencing my demo issue in the footer below. Another nice feature of
github and other cloud based VCS systems it that it automatically links
to the demo issue I created using the hash syntax, linking this commit
and the issue in our repository.

Issue #1